requirements.txt
lightgbm==4.3.0
sentence-transformers==3.0.1
onnxruntime==1.19.2
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
faiss-cpu==1.8.0
torch==2.4.1
transformers==4.44.2
hydra-core==1.3.2
omegaconf==2.3.0
pytest==8.3.2
datasets==3.0.1
pyarrow==17.0.0
langdetect==1.0.9
rank-bm25==0.2.2
zenml==0.62.0
onnxmltools==1.12.0
onnxconverter-common==1.14.0
hummingbird-ml==0.4.10
